Refactor decimation into free function
